backend/myproject/requirements.txt
django
djangorestframework 
django-cors-headers
pandas
matplotlib
sklearn
pytesseract
pymongo
nltk